requirement.txt
transformers==4.6.1
seqeval==0.0.12
pytorch-crf==0.7.2
sentencepiece
tensorboard
numpy
tqdm